requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[argon2]
argon2-cffi
email-validator
python-multipart
python-dotenv